pushed helper funtion to util file
