+++ src/functions/vacation-request/update-vacation-request/handler.ts
@@ -5,8 +5,8 @@ import { CreateKeycloakApiService } from "src/database/services/keycloak-api-ser
 import type { VacationRequest } from "src/generated/homeLambdasModels/model/vacationRequest";
 import {
   getContractedWeek,
-  splitVacationDaysByYear,
-  updateRemainingVacationDays
+  getLatestStatus,
+  processVacationApproval
 } from "src/libs/vacation-utils";
 import {
   notifyAdminsVacationSubmittedAll,
@@ -72,10 +72,8 @@ const updateVacationRequestHandler: ValidatedEventAPIGatewayProxyEvent<
     };
   }
 
-  const existingLatestStatus = Array.isArray(existingVacationRequest.status)
-    ? existingVacationRequest.status.at(-1)?.status
-    : "UNKNOWN";
-  const newLatestStatus = Array.isArray(status) ? status.at(-1)?.status : "UNKNOWN";
+  const existingLatestStatus = getLatestStatus(existingVacationRequest.status);
+  const newLatestStatus = getLatestStatus(status);
 
   const statusChanged = existingLatestStatus !== newLatestStatus;
   const draftStatusChanged = existingVacationRequest.draft !== draft;
@@ -103,20 +101,16 @@ const updateVacationRequestHandler: ValidatedEventAPIGatewayProxyEvent<
   try {
     const updatedVacationRequest =
       await vacationRequestService.updateVacationRequest(vacationRequestUpdates);
-    if (currentStatus === "APPROVED") {
-      const daysByYear = splitVacationDaysByYear(startDate, endDate, contractedWeek);
-      for (const [year, daysInYear] of Object.entries(daysByYear)) {
-        const success = await updateRemainingVacationDays(userId, daysInYear, currentStatus, year);
-        if (!success) {
-          return {
-            statusCode: 400,
-            body: JSON.stringify({
-              message: `Cannot approve request: user does not have enough remaining vacation days for ${year}.`
-            })
-          };
-        }
-      }
-    }
+
+    const approvalError = await processVacationApproval(
+      userId,
+      startDate,
+      endDate,
+      contractedWeek,
+      currentStatus
+    );
+    if (approvalError) return approvalError;
+
     if (statusChanged) {
       await notifyUserVacationStatusUpdatedAll({
         email: userDetails.email,
